Remove marker-related props from MapVEuMap
