+++ packages/libs/components/src/map/SemanticMarkers.tsx
@@ -1,4 +1,11 @@
-import { ReactElement, useEffect, useState, cloneElement } from 'react';
+import {
+  ReactElement,
+  useEffect,
+  useState,
+  cloneElement,
+  useMemo,
+  useCallback,
+} from 'react';
 import {
   MarkerProps,
   BoundsViewport,
@@ -10,7 +17,7 @@ import { useMap } from 'react-leaflet';
 import { LatLngBounds } from 'leaflet';
 import { debounce } from 'lodash';
 
-interface SemanticMarkersProps {
+export interface SemanticMarkersProps {
   onBoundsChanged: (bvp: BoundsViewport) => void;
   markers: Array<ReactElement<BoundsDriftMarkerProps>>;
   recenterMarkers?: boolean;
@@ -19,6 +26,10 @@ interface SemanticMarkersProps {
     duration: number;
     animationFunction: AnimationFunction;
   } | null;
+  /** Whether to zoom and pan map to center on markers */
+  flyToMarkers?: boolean;
+  /** How long (in ms) after rendering to wait before flying to markers */
+  flyToMarkersDelay?: number;
 }
 
 /**
@@ -32,9 +43,15 @@ export default function SemanticMarkers({
   markers,
   animation,
   recenterMarkers = true,
+  flyToMarkers,
+  flyToMarkersDelay,
 }: SemanticMarkersProps) {
   // react-leaflet v3
   const map = useMap();
+
+  markers = useMemo(() => {
+    return markers.map((marker) => cloneElement(marker, { showPopup: true }));
+  }, [markers]);
 
   const [prevMarkers, setPrevMarkers] =
     useState<ReactElement<BoundsDriftMarkerProps>[]>(markers);
@@ -81,6 +98,7 @@ export default function SemanticMarkers({
   useEffect(() => {
     let recenteredMarkers = false;
     if (recenterMarkers && bounds) {
+      // eslint-disable-next-line react-hooks/exhaustive-deps
       markers = markers.map((marker) => {
         let { lat, lng } = marker.props.position;
         let {
@@ -145,9 +163,9 @@ export default function SemanticMarkers({
      **/
     let timeoutVariable: NodeJS.Timeout;
 
-    if (zoomType == 'in') {
+    if (zoomType === 'in') {
       setConsolidatedMarkers([...markers]);
-    } else if (zoomType == 'out') {
+    } else if (zoomType === 'out') {
       /** If we are zooming out then remove the old markers after they finish animating. **/
       timeoutVariable = setTimeout(
         () => {
@@ -158,7 +176,9 @@ export default function SemanticMarkers({
     }
 
     return () => clearTimeout(timeoutVariable);
-  }, [zoomType, markers]);
+  }, [zoomType, markers, animation]);
+
+  useFlyToMarkers({ markers, flyToMarkers, flyToMarkersDelay });
 
   return <>{consolidatedMarkers}</>;
 }
@@ -217,4 +237,98 @@ function constrainLongitudeToMainWorld({
     southWest: { lat: south, lng: newWest },
     northEast: { lat: north, lng: newEast },
   };
+}
+
+// for flyTo
+interface PerformFlyToMarkersProps {
+  /* markers */
+  markers: ReactElement<BoundsDriftMarkerProps>[];
+  /** Whether to zoom and pan map to center on markers */
+  flyToMarkers?: boolean;
+  /** How long (in ms) after rendering to wait before flying to markers */
+  flyToMarkersDelay?: number;
+}
+
+// component to implement flyTo functionality
+function useFlyToMarkers(props: PerformFlyToMarkersProps) {
+  const { markers, flyToMarkers, flyToMarkersDelay } = props;
+
+  // instead of using useRef() to the map in v2, useMap() should be used instead in v3
+  const map = useMap();
+
+  const markersBounds = useMemo(() => {
+    return computeMarkersBounds(markers);
+  }, [markers]);
+
+  const performFlyToMarkers = useCallback(() => {
+    if (markersBounds) {
+      const boundingBox = computeBoundingBox(markersBounds);
+      if (boundingBox) map.fitBounds(boundingBox);
+    }
+  }, [markersBounds, map]);
+
+  useEffect(() => {
+    const asyncEffect = async () => {
+      if (flyToMarkersDelay)
+        await new Promise((resolve) => setTimeout(resolve, flyToMarkersDelay));
+      performFlyToMarkers();
+    };
+
+    if (flyToMarkers && markers.length > 0) asyncEffect();
+  }, [markers, flyToMarkers, flyToMarkersDelay, performFlyToMarkers]);
+
+  return null;
+}
+
+// compute bounding box
+function computeBoundingBox(markersBounds: Bounds | null) {
+  if (markersBounds) {
+    const ne = markersBounds.northEast;
+    const sw = markersBounds.southWest;
+
+    const bufferFactor = 0.1;
+    const latBuffer = (ne.lat - sw.lat) * bufferFactor;
+    const lngBuffer = (ne.lng - sw.lng) * bufferFactor;
+
+    const boundingBox = new LatLngBounds([
+      [sw.lat - latBuffer, sw.lng - lngBuffer],
+      [ne.lat + latBuffer, ne.lng + lngBuffer],
+    ]);
+
+    return boundingBox;
+  } else {
+    return undefined;
+  }
+}
+
+// compute markers bounds
+function computeMarkersBounds(markers: ReactElement<BoundsDriftMarkerProps>[]) {
+  if (markers) {
+    let [minLat, maxLat, minLng, maxLng] = [90, -90, 180, -180];
+
+    for (const marker of markers) {
+      const bounds = marker.props.bounds;
+      const ne = bounds.northEast;
+      const sw = bounds.southWest;
+
+      if (ne.lat > maxLat) maxLat = ne.lat;
+      if (ne.lat < minLat) minLat = ne.lat;
+
+      if (ne.lng > maxLng) maxLng = ne.lng;
+      if (ne.lng < minLng) minLng = ne.lng;
+
+      if (sw.lat > maxLat) maxLat = sw.lat;
+      if (sw.lat < minLat) minLat = sw.lat;
+
+      if (sw.lng > maxLng) maxLng = sw.lng;
+      if (sw.lng < minLng) minLng = sw.lng;
+    }
+
+    return {
+      southWest: { lat: minLat, lng: minLng },
+      northEast: { lat: maxLat, lng: maxLng },
+    };
+  } else {
+    return null;
+  }
 }
