+++ packages/libs/components/src/stories/SidebarResize.stories.tsx
@@ -33,6 +33,7 @@ import MapVEuMap, { MapVEuMapProps } from '../map/MapVEuMap';
 // import Geohash from 'latlon-geohash';
 // import {DriftMarker} from "leaflet-drift-marker";
 import geohashAnimation from '../map/animation_functions/geohash';
+import SemanticMarkers from '../map/SemanticMarkers';
 
 export default {
   title: 'Sidebar/Sidebar in map',
@@ -227,13 +228,13 @@ export const SidebarResize: Story<MapVEuMapProps> = (args) => {
           <p>Help/Tutorial</p>
         </Tab>
       </Sidebar>
-      <MapVEuMap
-        {...args}
-        viewport={viewport}
-        onBoundsChanged={handleViewportChanged}
-        markers={markerElements}
-        animation={defaultAnimation}
-      />
+      <MapVEuMap {...args} viewport={viewport}>
+        <SemanticMarkers
+          onBoundsChanged={handleViewportChanged}
+          markers={markerElements}
+          animation={defaultAnimation}
+        />
+      </MapVEuMap>
     </>
   );
 };
